Revamped how email filters are saved and managed
