+++ src/components/etDataGrid/EtDataGrid.vue
@@ -6,6 +6,7 @@
                 :columns="columns"
                 :rowInfo="rowInfo"
                 :bulk-methods="bulkMethods"
+                :onFilterSave="onFilterSave"
             />
             <EtDataGridContent
                 :columns="columns"
@@ -39,16 +40,17 @@ import type {
     BulkMethod,
     CellWidthProvide,
     CheckedProvide,
-    FilterDefinition,
-    FilterObject,
-    FilterSavingProvide,
-    FiltersProvide,
     PaginationProvide,
     RowObject,
     RowVersionProvider,
     SortingObject,
     SortingProvide
 } from "./interfaces/DataGridMethods";
+import type {
+    FilterDefinition,
+    FilterObject,
+    FiltersProvide
+} from "./interfaces/DataGridFilters";
 import { useChecked } from "./composables/useChecked";
 import { useSorting } from "./composables/useSorting";
 import { useFilters } from "./composables/useFilters";
@@ -60,7 +62,6 @@ import { useRouter, useRoute } from "vue-router";
 import { useCellWidth } from "./composables/useCellWidth";
 import { useRowVersion } from "./composables/useRowVersion";
 import { assignToPath } from "./services/DataGridCellHelpers";
-import { useFilterSaving } from "./composables/useFilterSaving";
 
 interface IDataGridCriteria {
     sorting: SortingObject;
@@ -116,15 +117,17 @@ const props = defineProps({
     isMultiSorting: {
         type: Boolean,
         default: false
+    },
+    onFilterSave: {
+        type: Function as PropType<
+            (name: string, filtersObj: FilterObject) => void | null
+        >,
+        default: null
     }
 });
 
 const emit = defineEmits<{
     (e: "checked", rows: RowObject[]): void;
-    (
-        e: "mounted",
-        ctx: { filterSaving: UnwrapNestedRefs<FilterSavingProvide> | null }
-    ): void;
 }>();
 
 const rows = ref<RowObject[]>([]);
@@ -134,10 +137,6 @@ const checkedRows = useChecked<RowObject>(props.rowInfo, () => rows.value);
 const sorting = useSorting<RowObject>(props.isMultiSorting);
 sorting.reset(props.columns);
 const filters = useFilters<RowObject>(() => props.filters);
-let filterSaving: UnwrapNestedRefs<FilterSavingProvide> | null = null;
-if (props.name) {
-    filterSaving = useFilterSaving(props.name, filters);
-}
 const cellWidth = useCellWidth();
 const pagination = usePagination();
 const rowVersion = useRowVersion<RowObject>(props.rowInfo?.idKey || "guid");
@@ -237,12 +236,6 @@ function searchData() {
 provide<CheckedProvide>("checkedRows", checkedRows);
 provide<SortingProvide>("sorting", sorting);
 provide<FiltersProvide>("filters", filters);
-if (filterSaving) {
-    provide<UnwrapNestedRefs<FilterSavingProvide>>(
-        "filterSaving",
-        filterSaving
-    );
-}
 provide<PaginationProvide>("pagination", pagination);
 provide<Ref<boolean>>("isLoading", isLoading);
 provide<() => void>("searchData", searchData);
@@ -308,10 +301,6 @@ const container = ref(null);
 onMounted(() => {
     calculateMaxCellWidth();
     window.addEventListener("resize", calculateMaxCellWidth);
-
-    emit("mounted", {
-        filterSaving: filterSaving
-    });
 });
 
 onBeforeUnmount(() => {
